+++ rag/src/application/risk_report_parser.py
@@ -1,0 +1,24 @@
+from __future__ import annotations
+
+import json
+import logging
+from typing import Any, Dict
+
+logger = logging.getLogger(__name__)
+
+
+class RiskReportParser:
+    # LLM 출력(JSON 문자열)을 파싱하는 전용 파서
+    def parse(self, raw: str) -> Dict[str, Any]:
+        try:
+            return json.loads(raw)
+        except json.JSONDecodeError:
+            pass
+        start = raw.find("{")
+        end = raw.rfind("}")
+        if start != -1 and end != -1 and end > start:
+            try:
+                return json.loads(raw[start : end + 1])
+            except json.JSONDecodeError:
+                logger.warning("Failed to parse JSON from LLM output.")
+        return {}
